refactor: rename helpers to browserUtils
