+++ packages/new-polymath-issuer/src/state/sagas/__tests__/wallet.ts
@@ -1,4 +1,4 @@
-import { onAddressChange } from '@polymathnetwork/sdk';
+import { browserUtils } from '@polymathnetwork/sdk';
 import { MockedStore } from '~/testUtils/helpers';
 
 // TODO @RafaelVidaurre: Finish these tests
@@ -7,10 +7,12 @@ jest.mock('@polymathnetwork/sdk', () => {
   const original = require.requireActual('@polymathnetwork/sdk');
   return {
     ...original,
-    onAddressChange: jest.fn().mockImplementation(() => {
-      return () => {};
-    }),
-    getNetworkId: jest.fn(),
+    browserUtils: {
+      onAddressChange: jest.fn().mockImplementation(() => {
+        return () => {};
+      }),
+      getNetworkId: jest.fn(),
+    },
   };
 });
 
@@ -23,7 +25,7 @@ describe('wallet sagas', () => {
 
   describe('on address changes', () => {
     test.skip('calls setWallet action when wallet address is changed', () => {
-      (onAddressChange as any).mockImplementation(() => {});
+      (browserUtils.onAddressChange as any).mockImplementation(() => {});
     });
   });
 });
