+++ packages/new-polymath-sdk/src/browserUtils.ts
