+++ packages/new-polymath-sdk/src/index.ts
@@ -1,3 +1,5 @@
+import * as browserUtils from './browserUtils';
+
 export * from './classes';
-export * from './helpers';
 export * from './types';
+export { browserUtils };
